test(abstract-utxo): move BIP-322 Proof tests
Issue: BTC-2650
